pyproject.toml
[tool.poetry]
name = "dadn-chill"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "dadn_chill"}]

[tool.poetry.dependencies]
python = "^3.8"
scikit-learn = "^1.2.2"
paho-mqtt = "^1.6.1"
python-dotenv = "^1.0.0"
pandas = "^2.0.0"
requests = "^2.28.2"


[tool.poetry.group.ml.dependencies]
scikit-learn = "^1.2.2"
pandas = "^2.0.0"
jupyter = "^1.0.0"
matplotlib = "^3.7.1"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
